// funnel-flow-studio/src/types/funnel.ts
export interface TextStyles {
  fontSize?: number;
  fontWeight?: 300 | 400 | 500 | 600 | 700 | 800;
  fontStyle?: 'normal' | 'italic';
  textDecoration?: 'none' | 'underline' | 'line-through';
  color?: string;
  textGradient?: string;
  textAlign?: 'left' | 'center' | 'right';
  lineHeight?: number;
  letterSpacing?: number;
  linkUrl?: string;
  linkTarget?: '_self' | '_blank';
}